Add helper to get randpom port in tests

+++ config/config.go
@@ -88,7 +88,7 @@ type mutableNodeConfig interface {
 	OverrideNodeSpecificValues(
 		federationNodes map[string]FederationNode,
 		gossipPeers map[string]GossipPeer,
-		gossipListenPort uint16,
+		gossipListenPort int,
 		nodePublicKey primitives.Ed25519PublicKey,
 		nodePrivateKey primitives.Ed25519PrivateKey,
 		constantConsensusLeader primitives.Ed25519PublicKey,
